ditch redundant file to keep code in single file for demo purposes
